VideoProfile fixed, readme fixed

+++ com.vzaar.api/Vzaar.cs
@@ -114,7 +114,7 @@ namespace com.vzaar.api
         }
 
         /// <summary>
-        /// vzaar uses the oEmbed open standard for allowing 3rd parties to integrated with the vzaar. 
+        /// vzaar uses the oEmbed open standard for allowing 3rd parties to integrated with the vzaar.
         /// </summary>
         /// <param name="videoId"></param>
         /// <returns></returns>
@@ -184,7 +184,7 @@ namespace com.vzaar.api
                     }
                 };
             }
-            
+
             return result;
         }
 
@@ -437,7 +437,7 @@ namespace com.vzaar.api
             if (query.replaceId != "")
                 data += "<replace_id>" + query.replaceId + "</replace_id>";
             data += "<guid>" + query.guid + "</guid><title>" + query.title + "</title><description>" + query.description + "</description><labels>";
-            data += String.Join( ",", query.labels ) + "</labels><profile>" + query.profile + "</profile>";
+            data += String.Join( ",", query.labels ) + "</labels><profile>" + (int)query.profile + "</profile>";
             if (query.transcode)
                 data += "<transcoding>true</transcoding>";
             data += "</video> </vzaar-api>";
@@ -608,7 +608,7 @@ namespace com.vzaar.api
                                 "</encoding_params>" +
                             "</link_upload>" +
                        "</vzaar-api>";
-            
+
             var response = executeRequest(url, "POST", data);
 
             var doc = new XmlDocument();
